+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 ## Unreleased
 
 - fire `input` event when value is changed on input, select or textarea elements
+- `invalid` events are now correctly being fired per input element when validating form rather than only the first one
 
 ## 11.15.0